fix: --longtests not available (#200)
